port packing from wasm-pack to trunk
