--- run.sh
@@ -1,2 +1,0 @@
-set -e
-bash build.sh && python -m http.server --directory publish
